refactor: move to AntennaService
